Bunker/requirements.txt
# Install with: `pip install -r requirements.txt`

cssselect
django
flake8
lxml
mock
pbr
pyquery
djangorestframework
corsheaders
